Update v0.6.7
Updated the 5js to v1.10.0
Updated the p5.sound.js (not in use)
Updated the documentation for easier setup,
Added comments and jsdocs for easier setup
Made easy to change resolution
Made easy to change frameRates (still not FPS independent tho).
updated the side menu a little bit to make it look good.